api/requirements.txt
# Dependências essenciais para a API
fastapi==0.110.1
uvicorn[standard]==0.25.0
python-dotenv>=1.0.1
motor==3.3.1
pydantic>=2.6.4
tzdata>=2024.2

# Para desenvolvimento/testes (opcional)
pytest>=8.0.0
